+++ .gitignore
@@ -1,0 +1,3 @@
+*.csv
+101501_example_SAFE.jl
+old_planet_fitting
