+++ src/general_functions.jl
@@ -3,14 +3,15 @@
 using LinearAlgebra
 using Random
 using Statistics
+using Unitful
 
 """
 Nyquist frequency is half of the sampling rate of a discrete signal processing system
 (https://en.wikipedia.org/wiki/Nyquist_frequency)
 divide by another factor of 4 for uneven spacing
 """
-nyquist_frequency(time_span::Real, n_meas::Integer; nyquist_factor::Real=1) = n_meas / time_span / 2 * nyquist_factor
-function nyquist_frequency(times::Vector{T}; nyquist_factor::Real=1) where {T<:Real}
+nyquist_frequency(time_span::Union{Real,Quantity}, n_meas::Integer; nyquist_factor::Real=1) = n_meas / time_span / 2 * nyquist_factor
+function nyquist_frequency(times::Vector{T}; nyquist_factor::Real=1) where {T<:Union{Real,Quantity}}
     time_span = times[end] - times[1]
     return nyquist_frequency(time_span, length(times), nyquist_factor=nyquist_factor)
 end
@@ -21,7 +22,7 @@ uneven_nyquist_frequency(times; nyquist_factor=5) = nyquist_frequency(times; nyq
 shamelessly crimped from JuliaAstro.jl
 used to calculate range of frequencies to look at in a periodogram
 """
-function autofrequency(times::Vector{T} where {T<:Real};
+function autofrequency(times::Vector{T} where {T<:Union{Real,Quantity}};
                        samples_per_peak::Integer=5,
                        nyquist_factor::Integer=5,
                        minimum_frequency::Real=NaN,
